+++ CHANGELOG.md
@@ -11,6 +11,58 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add `AggregateError` inspired from TC39.
 - Add `decode` method to `Type` module to help wrapping `io-ts` `Errors` into an `Error` subclass.
 - Add unit tests for the `set` function of the `Redis` module.
+- Add `Has` module and enhance `ReaderTaskEither` for smart dependencies management (inspired by [Effect-TS](https://www.matechs.com/open-source)):
+
+  ```typescript
+  // Foo.ts
+  import { TaskEither } from 'fp-ts/TaskEither'
+  import { $has, $readerTaskEither } from 'fortepiano'
+
+  export interface Foo {
+    bar(a: number): TaskEither<Error, string>
+  }
+
+  export const TagFoo = $has.tag<Foo>()
+
+  export const $foo = {
+    bar: $readerTaskEither.derivesTaskEither(TagFoo, 'bar'),
+  }
+  ```
+
+  ```typescript
+  // Bar.ts
+  import { IOEither } from 'fp-ts/IOEither'
+  import { $has, $readerTaskEither } from 'fortepiano'
+
+  export interface Bar {
+    (a: string): IOEither<Error, boolean>
+  }
+
+  export const TagBar = $has.tag<Bar>()
+
+  export const bar = $readerTaskEither.deriveIOEither(TagBar)
+  ```
+
+  ```typescript
+  // index.ts
+  import { $has } from 'fortepiano'
+  import { ioEither, readerTaskEither, taskEither } from 'fp-ts'
+  import { pipe } from 'fp-ts/function'
+  import { TagBar, bar } from './Bar'
+  import { $foo, TagFoo } from './Foo'
+
+  // const a: ReaderTaskEither<Has<Foo> & Has<Bar>, Error, boolean>
+  const a = pipe($foo.bar(42), readerTaskEither.chainW(bar))
+
+  // const b: TaskEither<Error, boolean>
+  const b = a(
+    // Let's mock our dependencies.
+    pipe(
+      $has.singleton(TagFoo, { bar: () => taskEither.of('foobar') }),
+      $has.upsertAt(TagBar, () => ioEither.of(true)),
+    ),
+  )
+  ```
 
 ### Changed